Update FMS-100 ( Save API for fixed and Bidding Indent)
Post Mapping:  “/api/indents/saveIndent”

Flow of event for this API:

Once the Schede trip or Place button is clicked(In the indent details page) a Post request with the following details will hit:

For fixed Indets:

Save the following information to the Subscribed Indents Table:

SO information

Route Information ( Chosen by the Transport manager).

Mark the Subscribed Indent from the SO table as Subscribed. (Since it must not be visible in New Indents)

For Bidding Indents:

Save the following information to the Subscribed Indents Table:

SO information

Bid chosen amount

Route Information

Mark the Subscribed Indent from the SO table as Subscribed.